Moving more things out of cache
